requirements.txt
rich
requests
openai
brave-search
